+++ nengo_spa/modules/tests/test_assoc_mem.py
@@ -62,7 +62,7 @@ def test_am_threshold(Simulator, plt, seed, rng):
         m.am = ThresholdingAssocMem(
             threshold=0.5, input_vocab=vocab, output_vocab=vocab2,
             function=filtered_step_fn, mapping='by-key')
-        m.stimulus = spa.Encode(input_func, vocab=vocab)
+        m.stimulus = spa.Transcode(input_func, output_vocab=vocab)
         spa.Actions(('am = stimulus',))
 
         in_p = nengo.Probe(m.am.input)
@@ -105,7 +105,7 @@ def test_am_wta(Simulator, plt, seed, rng):
     with spa.Network('model', seed=seed) as m:
         m.am = WTAAssocMem(
             threshold=0.3, input_vocab=vocab, function=filtered_step_fn)
-        m.stimulus = spa.Encode(input_func, vocab=vocab)
+        m.stimulus = spa.Transcode(input_func, output_vocab=vocab)
         spa.Actions(('am = stimulus',))
 
         in_p = nengo.Probe(m.am.input)
@@ -145,7 +145,7 @@ def test_am_default_output(Simulator, plt, seed, rng):
         m.am = ThresholdingAssocMem(threshold=0.5, input_vocab=vocab,
                                     function=filtered_step_fn)
         m.am.add_default_output('D', 0.5)
-        m.stimulus = spa.Encode(input_func, vocab=vocab)
+        m.stimulus = spa.Transcode(input_func, output_vocab=vocab)
         spa.Actions(('am = stimulus',))
 
         in_p = nengo.Probe(m.am.input)
@@ -191,8 +191,8 @@ def test_am_spa_keys_as_expressions(Simulator, plt, seed, rng):
             threshold=0.3, input_vocab=vocab_in, output_vocab=vocab_out,
             mapping=mapping)
 
-        model.inp = spa.Encode(
-            lambda t: 'A' if t < 0.1 else 'A*B', vocab=vocab_in)
+        model.inp = spa.Transcode(
+            lambda t: 'A' if t < 0.1 else 'A*B', output_vocab=vocab_in)
         spa.Actions(('am = inp',))
 
         in_p = nengo.Probe(model.am.input)